refactor: crate separation & dependency management (#162)
