+++ sbsp_app/Cargo.toml
@@ -19,28 +19,29 @@ name = "sbsp_app_lib"
 crate-type = ["staticlib", "cdylib", "rlib"]
 
 [build-dependencies]
-tauri-build = { version = "2", features = [] }
-sbsp_backend = { path = "../sbsp_backend", features = ["type_export"], default-features = false}
-sbsp_license = { path = "../sbsp_license", features = ["ts-rs"] }
+tauri-build.workspace = true
+sbsp_backend = { path = "../sbsp_backend", features = ["type_export"], default-features = false }
+sbsp_frontend_settings = { path = "../sbsp_frontend_settings", features = ["type_export"], default-features = false }
+sbsp_license = { path = "../sbsp_license", features = ["type_export"], default-features = false }
 
 [dependencies]
-tauri = { version = "2", features = [] }
-tauri-plugin-dialog = "2"
-tauri-plugin-log = { version = "2", features = ["colored"] }
-tauri-plugin-os = "2"
-tauri-plugin-opener = "2"
-humantime = "2.3.0"
+tauri.workspace = true
+tauri-plugin-dialog.workspace = true
+tauri-plugin-log.workspace = true
+tauri-plugin-os.workspace = true
+tauri-plugin-opener.workspace = true
+humantime.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+
 sbsp_backend = { path = "../sbsp_backend", features = ["apiserver"] }
+sbsp_frontend_settings = { path = "../sbsp_frontend_settings" }
 sbsp_license = { path = "../sbsp_license" }
-tokio = { version = "1", features = ["rt", "rt-multi-thread", "sync", "fs"] }
-log = "0.4"
-anyhow = "1.0"
-serde_json = "1.0"
-serde = { version = "1" }
-thiserror = "2"
-ts-rs = { version = "12" }
+tokio = { workspace = true, features = ["rt", "rt-multi-thread", "sync", "fs"] }
+log.workspace = true
+anyhow.workspace = true
 tower-http = { version = "0.6", features = ["fs"] }
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
-tauri-plugin-updater = "2"
-tauri-plugin-window-state = "2"
+tauri-plugin-updater.workspace = true
+tauri-plugin-window-state.workspace = true
