+++ sbsp_app/src/settings.rs
@@ -1,274 +1,64 @@
 // SPDX-License-Identifier: Elastic-2.0
 // Copyright (c) 2025 Keinsleif (https://github.com/Keinsleif)
 
-pub mod hotkey;
-pub mod manager;
+use std::path::{Path, PathBuf};
 
-use serde::{Deserialize, Serialize};
-use std::path::PathBuf;
-use ts_rs::TS;
+use sbsp_frontend_settings::{manager::SettingsManager, GlobalHostSettings};
+use sbsp_backend::BackendSettings;
+use tokio::sync::{RwLockReadGuard, watch};
 
-use hotkey::HotkeySettings;
-use sbsp_backend::{
-    BackendAudioSettings, BackendSettings,
-    model::cue::{
-        Cue, CueChain, CueColor, CueParam, FadeCueParam, LoadCueParam, PauseCueParam,
-        StartCueParam, StopCueParam, Uuid, WaitCueParam,
-        audio::{AudioCueParam, Decibels, Easing, FadeParam, SoundType},
-        group::{GroupCueParamBase, GroupMode},
-    },
-};
-
-#[derive(Serialize, Deserialize, Debug, Clone, Default, PartialEq, TS)]
-#[ts(export)]
-#[serde(rename_all = "camelCase", default)]
-pub struct GlobalHostSettings {
-    pub general: GeneralSettings,
-    #[serde(default)]
-    pub audio: AudioHardwareSettings,
-    pub appearance: AppearanceSettings,
-    pub hotkey: HotkeySettings,
-    pub template: TemplateSettings,
-    pub name_format: NameFormatSettings,
+pub struct GlobalSettingsManager {
+    inner: SettingsManager<GlobalHostSettings>,
+    settings_tx: watch::Sender<BackendSettings>,
 }
 
-impl From<&GlobalHostSettings> for BackendSettings {
-    fn from(from: &GlobalHostSettings) -> BackendSettings {
-        BackendSettings {
-            advance_cursor_when_go: from.general.advance_cursor_when_execute,
-            copy_assets_when_add: from.general.copy_assets_when_add,
-            audio: BackendAudioSettings {
-                device_id: from.audio.device_id.clone(),
-                channel_count: from.audio.channel_count,
-                sample_rate: from.audio.sample_rate,
-                buffer_size: from.audio.buffer_size,
-            },
-        }
+impl GlobalSettingsManager {
+    pub fn new(path: Option<PathBuf>) -> (Self, watch::Receiver<BackendSettings>) {
+        let inner = SettingsManager::new(path);
+        let (settings_tx, settings_rx) =
+            watch::channel(BackendSettings::from(&GlobalHostSettings::default()));
+        (Self { inner, settings_tx }, settings_rx)
     }
-}
 
-impl PartialEq<BackendSettings> for GlobalHostSettings {
-    fn eq(&self, other: &BackendSettings) -> bool {
-        if self.general.advance_cursor_when_execute == other.advance_cursor_when_go
-            && self.general.copy_assets_when_add == other.copy_assets_when_add
-        {
-            return true;
-        }
-        false
+    pub async fn set(&self, new_settings: GlobalHostSettings) {
+        self.settings_tx.send_modify(|b| *b = BackendSettings::from(&new_settings));
+        self.inner.set(new_settings).await;
     }
-}
 
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct GeneralSettings {
-    #[serde(alias = "advanceCursorWhenGo")]
-    pub advance_cursor_when_execute: bool,
-    pub lock_cursor_to_selection: bool,
-    pub copy_assets_when_add: bool,
-    pub seek_amount: f64,
-}
+    pub async fn import_from_file(&self, path: &Path) -> Result<GlobalHostSettings, anyhow::Error> {
+        let content = tokio::fs::read_to_string(path).await?;
+        let mut settings = tokio::task::spawn_blocking(move || serde_json::from_str::<GlobalHostSettings>(&content)).await??;
+        sanitize_audio(&mut settings);
+        self.set(settings.clone()).await;
+        self.save().await?;
 
-impl Default for GeneralSettings {
-    fn default() -> Self {
-        Self {
-            advance_cursor_when_execute: false,
-            lock_cursor_to_selection: true,
-            copy_assets_when_add: false,
-            seek_amount: 5.0,
-        }
+        log::info!("GlobalSettings imported from: {}", path.display());
+        Ok(settings)
     }
-}
 
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct AudioHardwareSettings {
-    pub device_id: Option<String>,
-    pub channel_count: Option<u16>,
-    pub sample_rate: Option<u32>,
-    pub buffer_size: Option<u32>,
-}
+    pub async fn export_to_file(&self, path: &Path) -> Result<(), anyhow::Error> {
+        let mut settings = self.inner.read().await.clone();
+        sanitize_audio(&mut settings);
+        let content =
+            tokio::task::spawn_blocking(move || serde_json::to_string_pretty(&settings)).await??;
 
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct AppearanceSettings {
-    pub language: Option<String>,
-    pub dark_mode: DarkMode,
-    pub hide_controls: bool,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default, TS)]
-#[serde(rename_all = "camelCase")]
-pub enum DarkMode {
-    #[default]
-    Dark,
-    Light,
-    System,
-}
-
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct TemplateSettings {
-    pub audio: Cue,
-    pub wait: Cue,
-    pub fade: Cue,
-    pub start: Cue,
-    pub stop: Cue,
-    pub pause: Cue,
-    pub load: Cue,
-    pub group: Cue,
-}
-
-impl Default for TemplateSettings {
-    fn default() -> Self {
-        Self {
-            audio: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Audio(AudioCueParam {
-                    target: PathBuf::new(),
-                    start_time: None,
-                    fade_in_param: None,
-                    end_time: None,
-                    fade_out_param: None,
-                    volume: Decibels::IDENTITY,
-                    pan: 0.0,
-                    repeat: false,
-                    sound_type: SoundType::Streaming,
-                    envelope: Vec::new(),
-                }),
-            },
-            wait: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Wait(WaitCueParam { duration: 5.0 }),
-            },
-            fade: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Fade(FadeCueParam {
-                    target: Uuid::nil(),
-                    volume: Decibels::IDENTITY,
-                    fade_param: FadeParam {
-                        duration: 3.0,
-                        easing: Easing::InOutPow(2.0),
-                    },
-                }),
-            },
-            start: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Start(StartCueParam {
-                    target: Uuid::nil(),
-                }),
-            },
-            stop: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Stop(StopCueParam {
-                    target: Uuid::nil(),
-                    hard: false,
-                }),
-            },
-            pause: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Pause(PauseCueParam {
-                    target: Uuid::nil(),
-                }),
-            },
-            load: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Load(LoadCueParam {
-                    target: Uuid::nil(),
-                }),
-            },
-            group: Cue {
-                id: Uuid::nil(),
-                number: "".to_string(),
-                name: None,
-                notes: "".to_string(),
-                color: CueColor::None,
-                pre_wait: 0.0,
-                chain: CueChain::DoNotChain,
-                parent_id: None,
-                params: CueParam::Group {
-                    base: GroupCueParamBase {
-                        mode: GroupMode::Playlist { repeat: true },
-                    },
-                    children: Vec::new(),
-                },
-            },
+        if let Some(parent) = path.parent() {
+            tokio::fs::create_dir_all(parent).await?;
         }
+        tokio::fs::write(path, content).await?;
+
+        log::info!("GlobalSettings saved to: {}", path.display());
+        Ok(())
     }
+
+    pub async fn read(&self) -> RwLockReadGuard<'_, GlobalHostSettings> { self.inner.read().await }
+    pub async fn load(&self) -> Result<GlobalHostSettings, anyhow::Error> { self.inner.load().await }
+    pub async fn save(&self) -> Result<(), anyhow::Error> { self.inner.save().await }
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, TS)]
-#[serde(rename_all = "camelCase", default)]
-pub struct NameFormatSettings {
-    pub audio: String,
-    pub wait: String,
-    pub fade: String,
-    pub start: String,
-    pub stop: String,
-    pub pause: String,
-    pub load: String,
-    pub group: String,
-}
-
-impl Default for NameFormatSettings {
-    fn default() -> Self {
-        Self {
-            audio: "{filename}".into(),
-            wait: "Wait {duration}".into(),
-            fade: "Fade {targetName}".into(),
-            start: "Start {targetName}".into(),
-            stop: "Stop {targetName}".into(),
-            pause: "Pause {targetName}".into(),
-            load: "Load {targetName}".into(),
-            group: "Group".into(),
-        }
-    }
+fn sanitize_audio(settings: &mut GlobalHostSettings) {
+    settings.audio.device_id = None;
+    settings.audio.channel_count = None;
+    settings.audio.sample_rate = None;
+    settings.audio.buffer_size = None;
 }
