+++ sbsp_backend/Cargo.toml
@@ -50,15 +50,15 @@ apiclient = [
     "dep:serde_json",
 ]
 cli = ["dep:env_logger", "dep:clap"]
-type_export = []
+type_export = ["dep:ts-rs"]
 
 [dependencies]
-serde = { version = "1.0", features = ["derive"] }
-uuid = { version = "1", features = ["serde", "v4"] }
+serde = { workspace = true, features = ["derive"] }
+uuid = { workspace = true, features = ["serde", "v4"] }
 indexmap = { version = "2", features = ["serde"] }
-log = { version = "0.4", optional = true }
-tokio = { version = "1", features = ["rt", "rt-multi-thread", "time", "macros", "sync", "fs", "signal"], optional = true }
-serde_json = { version = "1.0", optional = true }
+log = { workspace = true, optional = true }
+tokio = { workspace = true, features = ["rt", "rt-multi-thread", "time", "macros", "sync", "fs", "signal"], optional = true }
+serde_json = { workspace = true, optional = true }
 rodio = { version = "0.22.2", default-features = false, features = ["playback", "symphonia-all"], optional = true }
 rtrb = { version = "0.3.3", optional = true}
 symphonia = { version = "0.5.4", features = ["all"], optional = true }
@@ -68,7 +68,7 @@ async-recursion = { version = "1.1.1", optional = true }
 normpath = { version = "1.5.1", optional = true }
 
 env_logger = { version = "0.11.8", optional = true }
-clap = { version = "4.6.0", features = ["derive"], optional = true }
+clap = { workspace = true, optional = true }
 
 rand = { version = "0.10", optional = true }
 base64 = { version = "0.22.1", optional = true }
@@ -82,7 +82,7 @@ axum = { version = "0.8", default-features = false, features = ["tokio", "http1"
 tokio-tungstenite = { version = "*", optional = true }
 futures-util = { version = "0.3", default-features = false, features = ["sink"], optional = true }
 
-ts-rs = { version = "12", features = ["uuid-impl", "indexmap-impl"] }
+ts-rs = { workspace = true, features = ["uuid-impl", "indexmap-impl"], optional = true }
 
 [dev-dependencies]
 tempfile = "3.22.0"
