+++ sbsp_backend/src/lib.rs
@@ -73,15 +73,12 @@ pub struct FullShowState {
     pub show_state: ShowState,
 }
 
-#[cfg(feature = "backend")]
 #[derive(Default)]
 pub struct BackendSettings {
-    pub advance_cursor_when_go: bool,
     pub copy_assets_when_add: bool,
     pub audio: BackendAudioSettings,
 }
 
-#[cfg(feature = "backend")]
 #[derive(Default, Clone, PartialEq)]
 pub struct BackendAudioSettings {
     pub device_id: Option<String>,
