+++ sbsp_frontend_settings/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "sbsp_frontend_settings"
+description = "SBSP settings and manager collection for host/remote frontend"
+version = "0.1.0"
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+homepage.workspace = true
+repository.workspace = true
+publish.workspace = true
+
+[features]
+default = ["manager"]
+manager = ["dep:anyhow", "dep:log", "dep:tokio", "dep:serde_json"]
+type_export = ["dep:ts-rs", "sbsp_backend/type_export"]
+
+[dependencies]
+sbsp_backend = { path = "../sbsp_backend", default-features = false }
+serde.workspace = true
+
+anyhow = { workspace = true, optional = true }
+log = { workspace = true, optional = true }
+serde_json = { workspace = true, optional = true }
+tokio = { workspace = true, features = ["sync"], optional = true }
+
+ts-rs = { workspace = true, optional = true }
