+++ sbsp_frontend_settings/src/hotkey.rs
@@ -2,16 +2,17 @@
 // Copyright (c) 2025 Keinsleif (https://github.com/Keinsleif)
 
 use serde::{Deserialize, Serialize};
-use ts_rs::TS;
 
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default, TS)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, Default)]
+#[cfg_attr(feature = "type_export", derive(ts_rs::TS))]
 #[serde(rename_all = "camelCase")]
 pub struct HotkeySettings {
     playback: PlaybackHotkey,
     audio_action: AudioActionHotkey,
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, TS)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "type_export", derive(ts_rs::TS))]
 #[serde(rename_all = "camelCase", default)]
 pub struct PlaybackHotkey {
     #[serde(alias = "go")]
@@ -36,13 +37,14 @@ impl Default for PlaybackHotkey {
             resume_all: Some("]".to_string()),
             stop: Some("Backspace".to_string()),
             stop_all: Some("Escape".to_string()),
-            seek_forward: None,
-            seek_backward: None,
+            seek_forward: Some("ArrowRight".to_string()),
+            seek_backward: Some("ArrowLeft".to_string()),
         }
     }
 }
 
-#[derive(Serialize, Deserialize, Debug, Clone, PartialEq, TS)]
+#[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
+#[cfg_attr(feature = "type_export", derive(ts_rs::TS))]
 #[serde(rename_all = "camelCase", default)]
 pub struct AudioActionHotkey {
     pub toggle_repeat: Option<String>,
