+++ sbsp_license/Cargo.toml
@@ -14,15 +14,20 @@ path = "src/main.rs"
 required-features = ["clap"]
 
 [dependencies]
-ed25519-dalek = { version = "2.2.0", features = ["pem"]}
-base64 = "0.22.1"
-anyhow = "1"
-serde_json = "1"
-serde = { version = "1", features = ["derive"] }
-uuid = { version = "1", features = ["serde", "v4"] }
-clap = { version = "4.5.53", features = ["derive"], optional = true }
-ts-rs = { version = "12", features = ["uuid-impl"], optional = true }
+serde = { workspace = true, features = ["derive"] }
+uuid = { workspace = true, features = ["serde", "v4"] }
+
+anyhow = { workspace = true, optional = true }
+ed25519-dalek = { version = "3.0", features = ["pem"], optional = true }
+base64 = { version = "0.23", optional = true }
+serde_json = { workspace = true, optional = true }
+
+clap = { workspace = true, optional = true }
+
+ts-rs = { workspace = true, features = ["uuid-impl"], optional = true }
 
 [features]
+default = ["manager"]
+manager = ["dep:anyhow", "dep:ed25519-dalek", "dep:base64", "dep:serde_json"]
+type_export = ["dep:ts-rs"]
 clap = ["dep:clap"]
-ts-rs = ["dep:ts-rs"]
