+++ sbsp_license/src/data.rs
@@ -5,14 +5,14 @@ use serde::{Deserialize, Serialize};
 use uuid::Uuid;
 
 #[derive(Serialize, Deserialize, Debug, Clone, PartialEq)]
-#[cfg_attr(feature = "ts-rs", derive(ts_rs::TS))]
+#[cfg_attr(feature = "type_export", derive(ts_rs::TS))]
 pub enum LicenseEdition {
     Free,
     Pro,
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
-#[cfg_attr(feature = "ts-rs", derive(ts_rs::TS))]
+#[cfg_attr(feature = "type_export", derive(ts_rs::TS))]
 pub struct LicenseInformation {
     pub owner: String,
     pub edition: LicenseEdition,
@@ -21,7 +21,7 @@ pub struct LicenseInformation {
 }
 
 #[derive(Serialize, Deserialize, Debug, Clone)]
-#[cfg_attr(feature = "ts-rs", derive(ts_rs::TS))]
+#[cfg_attr(feature = "type_export", derive(ts_rs::TS))]
 pub struct LicenseFile {
     pub payload: LicenseInformation,
     pub signature: String,
