+++ sbsp_remote_app/Cargo.toml
@@ -19,24 +19,26 @@ name = "sbsp_remote_app_lib"
 crate-type = ["staticlib", "cdylib", "rlib"]
 
 [build-dependencies]
-tauri-build = { version = "2", features = [] }
-sbsp_backend = { path = "../sbsp_backend", features = ["type_export"], default-features = false}
+tauri-build.workspace = true
+sbsp_backend = { path = "../sbsp_backend", features = ["type_export"], default-features = false }
+sbsp_frontend_settings = { path = "../sbsp_frontend_settings", features = ["type_export"], default-features = false }
 
 [dependencies]
-tauri = { version = "2", features = [] }
-tauri-plugin-dialog = "2"
-tauri-plugin-log = { version = "2", features = ["colored"] }
-tauri-plugin-os = "2"
-tauri-plugin-opener = "2"
-humantime = "2.3.0"
+tauri.workspace = true
+tauri-plugin-dialog.workspace = true
+tauri-plugin-log.workspace = true
+tauri-plugin-os.workspace = true
+tauri-plugin-opener.workspace = true
+humantime.workspace = true
+serde.workspace = true
+serde_json.workspace = true
+
 sbsp_backend = { path = "../sbsp_backend", features = ["apiclient"] }
-tokio = { version = "1", features = ["rt", "rt-multi-thread", "sync"] }
-log = "0.4"
-anyhow = "1.0"
-serde_json = { version = "1.0" }
-serde = { version = "1" }
-ts-rs = { version = "12" }
+sbsp_frontend_settings = { path = "../sbsp_frontend_settings" }
+tokio = { workspace = true, features = ["rt", "rt-multi-thread", "sync"] }
+log.workspace = true
+anyhow.workspace = true
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
-tauri-plugin-updater = "2"
-tauri-plugin-window-state = "2"
+tauri-plugin-updater.workspace = true
+tauri-plugin-window-state.workspace = true
