+++ sbsp_remote_app/src/lib.rs
@@ -2,7 +2,6 @@
 // Copyright (c) 2025 Keinsleif (https://github.com/Keinsleif)
 
 mod command;
-mod settings;
 
 use std::time::SystemTime;
 
@@ -15,11 +14,12 @@ use sbsp_backend::{
     },
     event::BackendEvent,
 };
+use sbsp_frontend_settings::GlobalRemoteSettings;
 use tauri::{AppHandle, Emitter, Manager as _, ipc::Channel};
 use tauri_plugin_log::fern::colors::{Color, ColoredLevelConfig};
 use tokio::sync::{Mutex, RwLock, broadcast, mpsc};
 
-use crate::settings::manager::GlobalSettingsManager;
+pub type GlobalSettingsManager = sbsp_frontend_settings::manager::SettingsManager<GlobalRemoteSettings>;
 
 #[cfg(debug_assertions)]
 const LOG_LEVEL: LevelFilter = LevelFilter::Debug;
